Update Nice App

+++ LibraryApp/App.Models/PostCategory.cs
@@ -12,7 +12,7 @@ namespace App.Models
            
         }
 		
-		public Int16 PostId { get; set; }
+		public Int32 PostId { get; set; }
         public Int16 CategoryId { get; set; }
 		public virtual Post Post{ get; set; }
 		public virtual Category Category{ get; set; }
